static/scheme-finder/requirements.txt
flask
flask-pymongo
pymongo
python-dotenv
groq
werkzeug
